[lldb] Add support for custom char8_t types with different name
Currently for UTF-16 and UTF-32 types that are recognized by the DWARF encoding
tags without `char16_t` and `char32_t` names respectively are recognized, but
other names different than `char8_t` are not. To support the same behaviour of
the previous encodings, DWARF enconding tags should be taken to consideration
before fallbacking into the type name.

Signed-off-by: Luís Ferreira <contact@lsferreira.net>