[lldb] Move UpdateSymbolContextScopeForType
The UpdateSymbolContextScopeForType method is only used by ParseTypeFromDWARF
at the end and it represents a sequence to update the symbol context scope and
assign the given type to the DIEToType map.

Moving this sequence to DWARFASTParser decouples this logic from the
Clang-specific parser and aim other DWARF parsers to be implemented for other
languages other than C/C++/ObjC/ObjC++ (non-Clang DWARF parsers).

Signed-off-by: Luís Ferreira <contact@lsferreira.net>